backend/apps/movies_admin/pyproject.toml
[tool.poetry]
name = "django-movies"
version = "0.1.0"
description = "yandex practicum"
authors = ["Nikita Smirnov <nikita.smirnov@glasgow.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-environ = "^0.10.0"
django-split-settings = "^1.2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-jose = "^3.3.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pre-commit = "^3.2.2"
pycodestyle = "^2.10.0"
ipython = "^8.13.1"
django-extensions = "^3.2.1"
django-debug-toolbar = "3.4.0"
django-cors-headers = "^3.14.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
